feat: Save backtest results from optimization jobs
This change modifies the optimization job to save the full backtest result of the best trial to a file. This allows the frontend to display the results in the analysis/backtest tab.

It also fixes a bug in the scheduler API where `jsonify` was being used incorrectly in Flask-RESTful resources.